+++ errors/cause_test.go
@@ -19,15 +19,6 @@ func Test_IsRootCause(t *testing.T) {
 		assert.False(t, IsRootCause(err, ValidationErrors{}))
 	})
 
-	t.Run("given an error stack with errgo.NoteMask", func(t *testing.T) {
-		var err error
-		err = (&ValidationErrors{})
-		err = errgo.NoteMask(err, "biniou")
-
-		assert.True(t, IsRootCause(err, &ValidationErrors{}))
-		assert.False(t, IsRootCause(err, ValidationErrors{}))
-	})
-
 	t.Run("given an error stack with errors.Wrap", func(t *testing.T) {
 		var err error
 		err = (&ValidationErrors{})
@@ -73,14 +64,6 @@ func Test_IsRootCause(t *testing.T) {
 		assert.False(t, IsRootCause(err, ValidationErrors{}))
 	})
 
-	t.Run("given an error stack with NoteMask from ErrCtx", func(t *testing.T) {
-		var err error
-		err = (&ValidationErrors{})
-		err = NoteMask(context.Background(), err, "biniou")
-
-		assert.True(t, IsRootCause(err, &ValidationErrors{}))
-		assert.False(t, IsRootCause(err, ValidationErrors{}))
-	})
 }
 
 func Test_RootCause(t *testing.T) {
@@ -104,18 +87,6 @@ func Test_RootCause(t *testing.T) {
 			},
 		})
 		err = errgo.Notef(err, "pouet")
-
-		assert.Equal(t, "test=biniou", RootCause(err).Error())
-	})
-
-	t.Run("given an error stack with errgo.NoteMask", func(t *testing.T) {
-		var err error
-		err = (&ValidationErrors{
-			Errors: map[string][]string{
-				"test": {"biniou"},
-			},
-		})
-		err = errgo.NoteMask(err, "pouet")
 
 		assert.Equal(t, "test=biniou", RootCause(err).Error())
 	})
@@ -167,18 +138,6 @@ func Test_RootCause(t *testing.T) {
 
 		assert.Equal(t, "test=biniou", RootCause(err).Error())
 	})
-
-	t.Run("given an error stack with NoteMask from ErrCtx", func(t *testing.T) {
-		var err error
-		err = (&ValidationErrors{
-			Errors: map[string][]string{
-				"test": {"biniou"},
-			},
-		})
-		err = NoteMask(context.Background(), err, "pouet")
-
-		assert.Equal(t, "test=biniou", RootCause(err).Error())
-	})
 }
 
 func Test_UnwrapError(t *testing.T) {
@@ -213,18 +172,6 @@ func Test_UnwrapError(t *testing.T) {
 		}
 
 		assert.Equal(t, "test=biniou", lastErr.Error())
-	})
-
-	t.Run("given an error stack with NoteMask from ErrCtx", func(t *testing.T) {
-		var err error
-		err = (&ValidationErrors{
-			Errors: map[string][]string{
-				"test": {"biniou"},
-			},
-		})
-		err = NoteMask(context.Background(), err, "pouet")
-
-		assert.Equal(t, "pouet: test=biniou", UnwrapError(err).Error())
 	})
 
 	t.Run("given an error stack with Notef from ErrCtx", func(t *testing.T) {